+++ source/Examples/Resources/CMSiteMaintenance/CMSiteMaintenance_UpdateAppCatTablesTask_Enabled.ps1
@@ -1,0 +1,20 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to enable and set Update Application Catalog Tables task.
+#>
+
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMSiteMaintenance UpdateAppCatTable
+        {
+            SiteCode    = 'Lab'
+            TaskName    = 'Update Application Catalog Tables'
+            Enabled     = $true
+            RunInterval = 1380
+        }
+    }
+}