+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMPxeDistributionPoint Resource
 - Added CMPullDistributionPoint Resource
 - Added ConvertTo-AnyCimInstance to the ResourceHelper
+- Added CMSiteMaintenance Resource
 
 ### Changed